+++ src/pages/EventDetails/EventDetailContainer.tsx
@@ -76,11 +76,13 @@ const EventDetails: React.FC = () => {
     <HomeLayout>
       <div className="min-h-screen bg-white rounded-b-3xl shadow-2xl max-w-7xl mx-auto  ">
         
-        <div className="container mx-auto px-4 py-8 lg:py-16 bg-green-200 w-[90vw] rounded-3xl
+        {/* <div className="container mx-auto px-4 py-8 lg:py-16 bg-green-200 w-[90vw] rounded-3xl */}
+        <div className="container mx-auto px-4 py-8 lg:py-16 bg-[#FFFFFF] w-[90vw] rounded-3xl
               relative top-[-5rem] z-20 shadow-md">
           {/* <div className="lg:flex gap-8 lg:gap-16 items-start justify-center"> */}
           <div className="grid grid-cols-1 lg:grid-cols-2 gap-8 xl:gap-16 
-          items-start justify-center w-full bg-red-400">
+          items-start justify-center w-full">
+          {/* items-start justify-center w-full bg-red-400"> */}
             {/* Event Details Section */}
             <motion.div
               // className="space-y-6 grid lg:grid-cols-2 gap-4 items-start justify-center"
@@ -110,19 +112,19 @@ const EventDetails: React.FC = () => {
                 variants={fadeInUp}
               >
                 <motion.h2
-                  className="text-[40px] lg:text-4xl font-bold mb-4 font-redhat"
+                  className="text-[40px] lg:text-4xl font-bold mb-4 font-redhat red-hat-display"
                   variants={fadeInUp}
                 >
                   Xtasy Groove
                 </motion.h2>
 
-                <div className="space-y-3 mb-6 text-black">
+                <div className="space-y-3 mb-6 text-[#231F20] font-semibold red-hat-display">
                   <motion.div
                     className="flex items-center gap-3"
                     variants={fadeInUp}
                   >
                     <Calendar className="w-5 h-5 text-primary/80" />
-                    <span>12 Apr 2025</span>
+                    <span className="r">12 Apr 2025</span>
                   </motion.div>
                   <motion.div
                     className="flex items-center gap-3"
@@ -144,7 +146,7 @@ const EventDetails: React.FC = () => {
                   <h3 className="text-xl font-semibold mb-3 text-black">
                     About Event
                   </h3>
-                  <p className="text-black leading-relaxed">
+                  <p className="text-[#918F90] leading-relaxed font-medium red-hat-display">
                     Dive into the seductive heartbeat of the night – a world
                     where basslines flirt with your soul and every pulse of
                     light teases your wild side. Xtasy Groove isn't just a
@@ -161,16 +163,17 @@ const EventDetails: React.FC = () => {
 
             {/* Ticket Purchase Section */}
             <motion.div
-              className="w-full lg:w-[31rem] bg-faintPink rounded-2xl shadow-2xl 
+              className="w-full lg:w-[31rem] bg-gradient-to-b from-[#FFF2F4] from-0% via-[#FFF2F4] via-70% to-[#ffffff]/50 to-90%
+              rounded-2xl shadow-2xl 
               p-6 lg:p-8 sticky top-8"
               initial={{ opacity: 0, x: 30 }}
               animate={{ opacity: 1, x: 0 }}
               transition={{ delay: 0.4, duration: 0.8 }}
             >
-              <h3 className="text-2xl font-bold text-gray-900 mb-2">
+              <h3 className="text-2xl font-bold text-[#231F20] mb-2 red-hat-display">
                 Get Tickets
               </h3>
-              <p className="text-gray-600 mb-6">
+              <p className="text-[#979595] font-semibold mb-6 red-hat-display">
                 Which ticket type are you going for?
               </p>
 
@@ -275,7 +278,7 @@ const EventDetails: React.FC = () => {
                         {formatPrice(ticketPrices.xtasy)}
                       </p>
                     </div>
-                    <div className="text-red-600 font-semibold">Sold out</div>
+                    <div className="text-primary font-semibold">Sold out</div>
                   </div>
                 </motion.div>
               </div>
@@ -286,13 +289,13 @@ const EventDetails: React.FC = () => {
                   <span className="text-xl font-bold text-gray-900">
                     Total:
                   </span>
-                  <span className="text-2xl font-bold text-red-600">
+                  <span className="text-2xl font-bold text-primary">
                     {formatPrice(calculateTotal(tickets, ticketPrices))}
                   </span>
                 </div>
 
                 <motion.button
-                  className="w-full bg-red-600 hover:bg-red-700 text-white font-bold py-4 px-6 rounded-xl text-lg transition-colors shadow-lg"
+                  className="w-full bg-primary hover:bg-red-700 text-white font-bold py-4 px-6 rounded-xl text-lg transition-colors shadow-lg"
                   whileHover={{ scale: 1.02 }}
                   whileTap={{ scale: 0.98 }}
                   onClick={() => handleNext()}
