+++ src/pages/Profile/Profile_Picture.tsx
@@ -2,7 +2,7 @@ import { useRef, useState } from "react";
 import { motion } from "framer-motion";
 import { Camera } from "lucide-react";
 
-const ProfilePicture = () => {
+const ProfilePicture: React.FC = () => {
   const fileInputRef = useRef<HTMLInputElement>(null);
   const [profileImage, setProfileImage] = useState(
     "https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?w=96&h=96&fit=crop&crop=face"
