+++ src/pages/Profile/Profile_page_component.tsx
@@ -1,8 +1,7 @@
 import React, { useState } from "react";
 import { motion } from "framer-motion";
 import { Edit3, ArrowLeft } from "lucide-react";
-import ProfilePicture from "./Profile_Picture";
-// import RadioButton from "../../components/inputs/RadioButton";
+import ProfilePicture from "../Profile/Profile_Picture";
 
 const ProfileComponent = () => {
   const [formData, setFormData] = useState({
