+++ src/pages/Profile/index.tsx
@@ -1,5 +1,5 @@
 import HomeLayout from "../../components/layouts/HomeLayout";
-import Profile_layout from "./Profile_layout";
+import Profile_layout from "../Profile/Profile_layout";
 // import ProfileComponent from './Profile_page_component';
 
 const Profile = () => {
